@@ -24,6 +24,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -59,21 +60,23 @@ public class AuthenticationFilter extends UsernamePasswordAuthenticationFilter {
 	@Override
 	protected void successfulAuthentication(HttpServletRequest req, HttpServletResponse res, FilterChain chain,
 			Authentication auth) throws IOException, ServletException {
-
+    
 		String userName = ((User) auth.getPrincipal()).getUsername();
 		UserDto userDetails = usersService.getUserDetailsByEmail(userName);
 		String tokenSecret = environment.getProperty("token.secret");
 		byte[] secretKeyBytes = Base64.getEncoder().encode(tokenSecret.getBytes());
-		SecretKey secretKey = new SecretKeySpec(secretKeyBytes, SignatureAlgorithm.HS512.getJcaName());
-
+        SecretKey secretKey = Keys.hmacShaKeyFor(secretKeyBytes);  
+		 
 		Instant now = Instant.now();
 
 		String token = Jwts.builder()
 				.claim("scope", auth.getAuthorities())
-				.setSubject(userDetails.getUserId())
-				.setExpiration(
+				.subject(userDetails.getUserId())
+				.expiration(
 						Date.from(now.plusMillis(Long.parseLong(environment.getProperty("token.expiration_time")))))
-				.setIssuedAt(Date.from(now)).signWith(secretKey, SignatureAlgorithm.HS512).compact();
+				.issuedAt(Date.from(now))
+				.signWith(secretKey)
+				.compact();
 
 		res.addHeader("token", token);
 		res.addHeader("userId", userDetails.getUserId());
